Reimplement tf2xla ResizeBilinearOp using gather.
Prior to this change, the `ResizeBilinearOp` was marked slow and skipped during XLA autoclustering. This change unmarks it so it can be clustered.

This also includes a check that ensures that one and only one of `align_corners` and `half_pixel_centers` is true. The (false, false) case is considered broken, and the (true, true) case is illegal.

PiperOrigin-RevId: 432024167